refactor(BA-5005): move TestNotificationProcessing to unit tests
TestNotificationProcessing directly called NotificationProcessors.
process_notification.wait_for_complete() — a service-layer method — without
going through the HTTP API.  This belongs in tests/unit/, not tests/component/.

- Add tests/unit/manager/services/notification/test_notification_processing.py
  with mocked repository (no real DB, no SDK)
- Remove TestNotificationProcessing class from component test file
- Remove webhook_channel and notification_rule fixtures from component
  notification conftest.py (they were only used by the moved class)
- Clean up unused imports in test_notification.py
